+++ bluebird/api/resources/ic.py
@@ -7,6 +7,7 @@ from flask_restful import Resource, reqparse
 
 import bluebird.client as bb_client
 import bluebird.settings as settings
+from bluebird.api.resources.utils import wait_for_data
 
 PARSER = reqparse.RequestParser()
 PARSER.add_argument('filename', type=str, location='json', required=True)
@@ -53,5 +54,7 @@ class Ic(Resource):
 		else:
 			resp = jsonify(f'Error: Could not load scenario {fn_base}. Error was: {err}')
 			resp.status_code = 500
+
+		wait_for_data()
 
 		return resp