+++ bluebird/api/resources/scenario.py
@@ -9,6 +9,7 @@ from flask import jsonify
 from flask_restful import Resource, reqparse
 
 import bluebird.client as bb_client
+from bluebird.api.resources.utils import wait_for_data
 from bluebird.logging import store_local_scn
 
 _LOGGER = logging.getLogger('bluebird')
@@ -19,7 +20,7 @@ PARSER.add_argument('content', type=str, location='json', required=True, action=
 PARSER.add_argument('start_new', type=bool, location='json', required=False)
 PARSER.add_argument('start_dtmult', type=float, location='json', required=False)
 
-_SCN_RE = re.compile(r'\d{2}:\d{2}:\d{2}\s?>\s?.*')
+_SCN_RE = re.compile(r'\d{2}:\d{2}:\d{2}(\.\d{1,3})?\s?>\s?.*')
 
 
 def _validate_scenario(scn_lines):
@@ -89,5 +90,7 @@ class Scenario(Resource):
 		else:
 			resp = jsonify(f'Scenario {scn_base} uploaded')
 			resp.status_code = 201
+
+		wait_for_data()
 
 		return resp